feat (#28): [V5.4] Deleting a task by the creator only - add check if user has access to route of delete task with voter key

+++ src/Controller/TaskController.php
@@ -92,6 +92,8 @@ class TaskController extends AbstractController
         Task $task,
         EntityManagerInterface $entityManager
     ): Response {
+        $this->denyAccessUnlessGranted("TASK_DELETE", $task);
+
         $entityManager->remove($task);
         $entityManager->flush();
